Adding Config
This adds a configuration class and helpers, making it much easier to
structure projects with Nippy. Configuration files are searched for
by default in `project_root/config`, but these paths can be
overwritten. A Config instance will try to cascade, if found, a
`default` config file with an `env` config file.

* `Config` class handling loading, getting and setting.
* `init` helper function to get or instantiate a Config.
* `config` helper function to get values on default Config.

A fair amount of tests have been defined, and some have been written.